fix: correct Ascend 910 V2 SWA bounds and invalid-row clearing

+++ csrc/ascend910/flash_attn_npu/mha_fwd_kvcache.cpp
@@ -696,7 +696,8 @@ namespace SplitFuse {
                 kvStartOld = CeilDiv(kvSeqlenOld, MAX_KV_STACK_LEN);
                 kvSLoopNumTotalOld = CeilDiv(
                     AscendC::Std::max((int64_t)0,
-                        AscendC::Std::min(noSkipKvS, (int64_t)kvSeqlenOld)),
+                        AscendC::Std::min(static_cast<int64_t>(noSkipKvS),
+                            static_cast<int64_t>(kvSeqlenOld))),
                     (int64_t)MAX_KV_STACK_LEN);
                 kvLoopNumTotalNew = CeilDiv(kvNewSeqlen, MAX_KV_STACK_LEN);
                 kvSLoopNumTotal = kvSLoopNumTotalOld + kvLoopNumTotalNew;
